renames classes and created tools and usecases
