+++ main.py
@@ -1,4 +1,4 @@
-import classes
+import modules.generators as generators
 import os
 from config import country, job_type, root_dir, job_portal
 
@@ -8,5 +8,5 @@ position = input('Position: ')
 workdir = root_dir + country
 
 if os.name == 'nt':
-    wingenerator = classes.GenerateDocsWin()
+    wingenerator = generators.GenerateDocsWin()
     wingenerator.generate(workdir, company, job_type, position, job_portal)
